+++ include/hpp/fcl/hfield.h
@@ -37,11 +37,11 @@
 #ifndef HPP_FCL_HEIGHT_FIELD_H
 #define HPP_FCL_HEIGHT_FIELD_H
 
-#include <hpp/fcl/fwd.hh>
-#include <hpp/fcl/data_types.h>
-#include <hpp/fcl/collision_object.h>
-#include <hpp/fcl/BV/BV_node.h>
-#include <hpp/fcl/BVH/BVH_internal.h>
+#include "hpp/fcl/fwd.hh"
+#include "hpp/fcl/data_types.h"
+#include "hpp/fcl/collision_object.h"
+#include "hpp/fcl/BV/BV_node.h"
+#include "hpp/fcl/BVH/BVH_internal.h"
 
 #include <vector>
 
@@ -207,7 +207,7 @@ class HPP_FCL_DLLAPI HeightField : public CollisionGeometry {
   typedef CollisionGeometry Base;
 
   typedef HFNode<BV> Node;
-  typedef std::vector<Node, Eigen::aligned_allocator<Node> > BVS;
+  typedef std::vector<Node, Eigen::aligned_allocator<Node>> BVS;
 
   /// @brief Constructing an empty HeightField
   HeightField()
@@ -528,13 +528,13 @@ template <>
 NODE_TYPE HeightField<OBBRSS>::getNodeType() const;
 
 template <>
-NODE_TYPE HeightField<KDOP<16> >::getNodeType() const;
+NODE_TYPE HeightField<KDOP<16>>::getNodeType() const;
 
 template <>
-NODE_TYPE HeightField<KDOP<18> >::getNodeType() const;
+NODE_TYPE HeightField<KDOP<18>>::getNodeType() const;
 
 template <>
-NODE_TYPE HeightField<KDOP<24> >::getNodeType() const;
+NODE_TYPE HeightField<KDOP<24>>::getNodeType() const;
 
 /// @}