+++ include/hpp/fcl/shape/convex.h
@@ -38,7 +38,7 @@
 #ifndef HPP_FCL_SHAPE_CONVEX_H
 #define HPP_FCL_SHAPE_CONVEX_H
 
-#include <hpp/fcl/shape/geometric_shapes.h>
+#include "hpp/fcl/shape/geometric_shapes.h"
 
 namespace hpp {
 namespace fcl {
@@ -71,12 +71,6 @@ class Convex : public ConvexBase {
 
   ~Convex();
 
-  /// @brief An array of PolygonT object.
-  /// PolygonT should contains a list of vertices for each polygon,
-  /// in counter clockwise order.
-  std::shared_ptr<std::vector<PolygonT>> polygons;
-  unsigned int num_polygons;
-
   /// based on http://number-none.com/blow/inertia/bb_inertia.doc
   Matrix3f computeMomentofInertia() const;
 
@@ -99,8 +93,14 @@ class Convex : public ConvexBase {
            std::shared_ptr<std::vector<PolygonT>> polygons,
            unsigned int num_polygons);
 
-  ///  @brief Clone (deep copy).
+  /// @brief Clone (deep copy)
   virtual Convex<PolygonT>* clone() const;
+
+  /// @brief An array of PolygonT object.
+  /// PolygonT should contains a list of vertices for each polygon,
+  /// in counter clockwise order.
+  std::shared_ptr<std::vector<PolygonT>> polygons;
+  unsigned int num_polygons;
 
  protected:
   void fillNeighbors();
@@ -110,6 +110,6 @@ class Convex : public ConvexBase {
 
 }  // namespace hpp
 
-#include <hpp/fcl/shape/details/convex.hxx>
+#include "hpp/fcl/shape/details/convex.hxx"
 
 #endif