+++ python/distance.cc
@@ -38,7 +38,10 @@
 
 #include <hpp/fcl/fwd.hh>
 #include <hpp/fcl/distance.h>
+#include <hpp/fcl/serialization/collision_data.h>
 #include "deprecation.hh"
+
+#include "serializable.hh"
 
 #ifdef HPP_FCL_HAS_DOXYGEN_AUTODOC
 #include "doxygen_autodoc/functions.h"
@@ -51,7 +54,7 @@ using namespace hpp::fcl::python;
 
 namespace dv = doxygen::visitor;
 
-struct DistanceRequestWrapper {
+struct DistanceResultWrapper {
   static Vec3f getNearestPoint1(const DistanceResult& res) {
     return res.nearest_points[0];
   }
@@ -99,7 +102,8 @@ void exposeDistanceAPI() {
             doxygen::class_attrib_doc<DistanceRequest>("enable_nearest_points"))
         .DEF_RW_CLASS_ATTRIB(DistanceRequest, enable_signed_distance)
         .DEF_RW_CLASS_ATTRIB(DistanceRequest, rel_err)
-        .DEF_RW_CLASS_ATTRIB(DistanceRequest, abs_err);
+        .DEF_RW_CLASS_ATTRIB(DistanceRequest, abs_err)
+        .def(SerializableVisitor<DistanceRequest>());
   }
 
   if (!eigenpy::register_symbolic_link_to_registered_type<
@@ -115,9 +119,9 @@ void exposeDistanceAPI() {
         .DEF_RW_CLASS_ATTRIB(DistanceResult, min_distance)
         .DEF_RW_CLASS_ATTRIB(DistanceResult, normal)
         //.def_readwrite ("nearest_points", &DistanceResult::nearest_points)
-        .def("getNearestPoint1", &DistanceRequestWrapper::getNearestPoint1,
+        .def("getNearestPoint1", &DistanceResultWrapper::getNearestPoint1,
              doxygen::class_attrib_doc<DistanceResult>("nearest_points"))
-        .def("getNearestPoint2", &DistanceRequestWrapper::getNearestPoint2,
+        .def("getNearestPoint2", &DistanceResultWrapper::getNearestPoint2,
              doxygen::class_attrib_doc<DistanceResult>("nearest_points"))
         .DEF_RO_CLASS_ATTRIB(DistanceResult, nearest_points)
         .DEF_RO_CLASS_ATTRIB(DistanceResult, o1)
@@ -126,7 +130,8 @@ void exposeDistanceAPI() {
         .DEF_RW_CLASS_ATTRIB(DistanceResult, b2)
 
         .def("clear", &DistanceResult::clear,
-             doxygen::member_func_doc(&DistanceResult::clear));
+             doxygen::member_func_doc(&DistanceResult::clear))
+        .def(SerializableVisitor<DistanceResult>());
   }
 
   if (!eigenpy::register_symbolic_link_to_registered_type<